refactor(payment): 금액 BigDecimal -> Long 자료형 변경

+++ payment-service/src/main/java/com/rushcrew/payment_service/infrastructure/event/PaymentCompletedEvent.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 public record PaymentCompletedEvent(
         UUID paymentId,
         UUID orderId,
-        BigDecimal totalAmount,
+        Long totalAmount,
         String currency,
         LocalDateTime completedAt,
         String eventType
@@ -15,7 +15,7 @@ public record PaymentCompletedEvent(
     public static PaymentCompletedEvent of(
             UUID paymentId,
             UUID orderId,
-            BigDecimal totalAmount,
+            Long totalAmount,
             String currency
     ) {
         return new PaymentCompletedEvent(
